Tweak media view gradients.
