+++ Signal/src/ViewControllers/CropScaleImageViewController.swift
@@ -6,33 +6,6 @@ import Foundation
 import MediaPlayer
 import SignalMessaging
 
-class OWSLayerView: UIView {
-    let layoutCallback : (() -> Void)
-
-    required init(frame: CGRect, layoutCallback : @escaping () -> Void) {
-        self.layoutCallback = layoutCallback
-        super.init(frame: frame)
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        self.layoutCallback = {
-        }
-        super.init(coder: aDecoder)
-    }
-
-    override var bounds: CGRect {
-        didSet {
-            layoutCallback()
-        }
-    }
-
-    override var frame: CGRect {
-        didSet {
-            layoutCallback()
-        }
-    }
-}
-
 // This kind of view is tricky.  I've tried to organize things in the 
 // simplest possible way.
 //
